refactor: 랜딩 페이지 푸터 컴포넌트 분리
